+++ test/check.dart
@@ -6,8 +6,8 @@ import 'package:stack_trace/stack_trace.dart';
 void main() {
   try {
     final environment = Environment();
-    final template = environment
-        .fromString('{% filter lower|escape %}<HEHE>{% endfilter %}');
+    final template = environment.fromString(
+        '{% filter upper|replace(\'FOO\', \'foo\') %}foobar{% endfilter %}');
     print(template.nodes);
     print(template.render());
   } catch (error, trace) {