+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 0.4.0-dev.11
-- No dynamic template imports. Only single constant template path/name.
-- `Template.renderMap` deprecated.
-- `Template.render` accepts `Map<String, Object?>`.
+- no dynamic template imports, only single constant template path/name.
+- `Template.renderMap` now deprecated.
+- `Template.render` now accepts `Map<String, Object?>`.
+- `FileSystemLoader` moved from `package:jinja/jinja.dart` to `package:jinja/loaders.dart`.
 - new statements:
   - with
 - internal improvments